Draw player scores

+++ player.go
@@ -1,6 +1,6 @@
 package main
 
 type Player struct {
-	Score  int8
+	Score  int
 	Paddle Paddle
 }
